add FailCount and refactor
